chore(*): move to DDD properly
