+++ Micro.Auth.Api/UserData/Extensions/IpAddressExtension.cs
@@ -1,7 +1,7 @@
 using System.Net;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Micro.Auth.Api.Controllers.Extensions
+namespace Micro.Auth.Api.UserData.Extensions
 {
     public static class IpAddressExtension
     {
