+++ Micro.Auth.Api/Users/ViewModels/UsersViewModel.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Micro.Auth.Api.Controllers.ViewModels
+namespace Micro.Auth.Api.Users.ViewModels
 {
     public class CreateUserRequest
     {
